feat: 适配ESP32-S3并修复引脚与驱动兼容性
- 将不可用引脚(GPIO22/23/25)重映射到S3安全引脚(GPIO42/47/48)
- 新增本地mdns和json组件以支持离线构建
- 迁移ADC驱动至oneshot模式，修复电池采样
- 修复OTA/wifi_manager/display等多处API兼容性问题
- 配置USB-Serial-JTAG控制台并抑制I2C弃用警告
- 修复状态栏格式化警告及触摸屏校准类型错误
- 将大配置结构体移出栈避免栈溢出
- 添加build.ps1一键构建脚本并更新引脚文档

+++ firmware/main/display/status_bar.c
@@ -53,7 +53,7 @@ void status_bar_draw(void)
         uint32_t up_ms = (uint32_t)(esp_timer_get_time() / 1000);
         uint32_t h = up_ms / 3600000;
         uint32_t m = (up_ms / 60000) % 60;
-        snprintf(time_str, sizeof(time_str), "U%02d:%02d", h, m);
+        snprintf(time_str, sizeof(time_str), "U%02d:%02d", (int)h, (int)m);
     }
 
     display_draw_status_bar(time_str, wifi_manager_get_rssi(), battery_get_percent(), s_left_w);
